+++ lib/main.dart
@@ -1,8 +1,9 @@
+// ignore_for_file: avoid_print
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'package:firebase_auth/firebase_auth.dart';
+import 'package:mywaylearn/firebase_options.dart';
 import 'package:mywaylearn/view/login_view.dart';
-import 'firebase_options.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
@@ -11,40 +12,18 @@ void main() {
     theme: ThemeData(
       primarySwatch: Colors.orange,
     ),
-    home: const RegisterView(),
-    // home: const LoginView(),
+    // home: const RegisterView(),
+    home: const HomePage(),
   ));
 }
 
-class RegisterView extends StatefulWidget {
-  const RegisterView({Key? key}) : super(key: key);
-
-  @override
-  State<RegisterView> createState() => _RegisterViewState();
-}
-
-class _RegisterViewState extends State<RegisterView> {
-  late final TextEditingController _email;
-  late final TextEditingController _password;
-
-  @override
-  void initState() {
-    _email = TextEditingController();
-    _password = TextEditingController();
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    _email.dispose();
-    _password.dispose();
-    super.dispose();
-  }
+class HomePage extends StatelessWidget {
+  const HomePage({ Key? key }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: const Text('Register')),
+      appBar: AppBar(title: const Text('Home')),
       body: FutureBuilder(
         future: Firebase.initializeApp(
           options: DefaultFirebaseOptions.currentPlatform,
@@ -52,48 +31,14 @@ class _RegisterViewState extends State<RegisterView> {
         builder: (context, snapshot) {
           switch (snapshot.connectionState) {
             case ConnectionState.done:
-              return Column(
-                children: [
-                  TextField(
-                    controller: _email,
-                    enableSuggestions: false,
-                    autocorrect: false,
-                    keyboardType: TextInputType.emailAddress,
-                    decoration: const InputDecoration(
-                      hintText: 'Enter your email here',
-                    ),
-                  ),
-                  TextField(
-                    controller: _password,
-                    obscureText: true,
-                    enableSuggestions: false,
-                    autocorrect: false,
-                    decoration: const InputDecoration(
-                      hintText: 'Enter your password here',
-                    ),
-                  ),
-                  TextButton(
-                    onPressed: () async {
-                      final email = _email.text;
-                      final password = _password.text;
-                      try{
-                        final userCredential = await FirebaseAuth.instance.createUserWithEmailAndPassword(email: email, password: password);
-                        print(userCredential);
-                      } on FirebaseAuthException catch (e) {
-                        print(e.code);
-                        if(e.code == 'weak-password') {
-                          print('To Weak Password.');
-                        } else if (e.code == 'email-already-in-use') {
-                          print('Email alredy in use. Please try with a different email.');
-                        } else if (e.code == 'invalid-email') {
-                          print('Invalid Email Address');
-                        }
-                      }
-                    },
-                    child: const Text('Register'),
-                  ),
-                ],
-              );
+              final user = FirebaseAuth.instance.currentUser;
+              final emailVerified = user?.emailVerified ?? false;
+              if (emailVerified) {
+                print('You are verified user');
+              } else {
+                print('You are not verified user, Please verify your email.');
+              }
+              return const Text('Done');
             default:
               return const Text('Loading.....!!!');
           }
